refactor: organize meme maker files
